+++ EPazar/EPazar/Controllers/OdemeController.cs
@@ -143,10 +143,10 @@ namespace EPazar.Controllers
             OdemeEntityleri.ViewSepet = viewSepets;
             OdemeEntityleri.Siparis = Siparis;
 
-            KullaniciAdresleri.UyeId = Siparis.UyeId;
+            KullaniciAdresleri.UyeId = (long)Siparis.UyeId;
             OdemeEntityleri.KullaniciAdresleri = await BusKullaniciAdresleri.PredicateAsync(KullaniciAdresleri);
 
-            KullaniciAdresleri.Id = Siparis.AdresId;
+            KullaniciAdresleri.Id = (long)Siparis.AdresId;
             OdemeEntityleri.SeciliKullaniciAdresi = await BusKullaniciAdresleri.FirstOrDefaultAsync(KullaniciAdresleri);
 
             return View(OdemeEntityleri);
@@ -216,10 +216,10 @@ namespace EPazar.Controllers
             OdemeEntityleri.ViewSepet = viewSepets;
             OdemeEntityleri.Siparis = Siparis;
 
-            KullaniciAdresleri.UyeId = Siparis.UyeId;
+            KullaniciAdresleri.UyeId = (long)Siparis.UyeId;
             OdemeEntityleri.KullaniciAdresleri = await BusKullaniciAdresleri.PredicateAsync(KullaniciAdresleri);
 
-            KullaniciAdresleri.Id = Siparis.AdresId;
+            KullaniciAdresleri.Id = (long)Siparis.AdresId;
             OdemeEntityleri.SeciliKullaniciAdresi = await BusKullaniciAdresleri.FirstOrDefaultAsync(KullaniciAdresleri);
 
             return View(OdemeEntityleri);
